Fix more balancing issues, Add missine Bluedog probes. Remove Bluedog INTERNAL antennae
Turns out a bunch of Bluedog antennae are internal with less than 5000
range which intend to be connectable. I've removed them as the most
likely use for them is to connect while on the launchpad, and you're
likely to attach antennae anyway.

I also added some forgotten probe cores and balanced both Bluedog and
Sounding Rockets.