Update OrbitalScience.cfg
This changes the DMagic antennae from the ridiculous 10 T to 280 Gm, and I changed the data Tx rate to 0.032 to bring it more in line with other antenna you'll have around this tech level. These give the player a real choice if they want to go really far away but only have limited Data throughput; they can now make that choice. They don't work in Atmosphere and can't relay since they are high-gain so they should be pretty balanced and provide a meaningful experience for people who want to send early probes to very distant bodies with the understanding they won't be able to transmit a lot of data quickly.

Removes the Soil Moisture Sensor as an Antennae.