Changes to implement a Profile importing scheme.
enable Profiles that are stored in any config file to be imported into
the loaded profile, this also allows modders to include their own
Supplies, Rules or Processes and not have to edit the default profile.

The code also checks a field supplied in the imported profile to
determine if the profile should be imported or not. This mechanism works
by searching the GameData directory for a matching directory that is
named in the moddir field of the profile.
Simply create a Profile node named KerbalismSupport and include the
necessary script to define the Supplies, Rules and Processes you need,
as you would with the default profile.
The KerbalismSupport profile node must have a modname field which is
used for logging and a moddir field which is used to decide if a mods
support profile should be loaded.
```
Profile
{
name = KerbalismSupport    // name must be KerbalismSupport
modname = FooBar Zoo       // Just a name for display purposes
moddir = FooBar/Zoo           // Could use Zoo if you wish

Supply
{
etc...
```

(cherry picked from commit 1aa431b47b05fa3d0cbf244904ad45aea675a8d7)