Humidity: Add new parameters to Rules
Added two parameters to the Rules definitions.

input_threshold    // when the input resource reaches this percentage of its max capacity, degeneration is triggered rather than the default of a zero resouce ammount [range 0 to 1].

output_only    // if true and an input resource exists, only monitor the input resource and do not consume it.