Greenhouse modules can now use combined WasteAtmosphere and CO2
Also when in breathable atmospheres the Greenhouse will consume CO2 from the external atmosphere and O2 will not be produced to simulate the O2 being released into the external atmosphere.

Greenhouses that do not specify WasteAtmosphere or CO2 as an Input will still function, this allows mods with greenhouses that are not using combined resources to still work.