Moved a bunch of Settings from the config file into the GUI (#196)
* Added first basic preferences screen

* GUI for game preferences

* Removed basic features from Preferences, there's no easy way to make them work.

The other settings (except comfort) are in effect, and removed from Settings.cfg

* Moved a bunch of other config settings to the GUI.

This *should* be working, not really tested yet.

* Updated changelog

* Shielding Efficiency format string fixed

* Moved Preferences.cs into System folder.
Added a setting to immediately transmit science, Fixes #59

* A binary! We have a binary! Hail and praise the gods!

* Death/Breakdown Penalty renamed to Death/Breakdown Reputation Penalty

* Tweaked the limits of radiation.

* Removed stock science dialog setting from the GUI.

This is being initialized when KSP starts. Will have a look at it on
how to do that at runtime (= on a per savegame basis), but until then
it's back in the settings.