Increase research speed in Laboratory if high level crew members are in it.
The laboratory will now give a 10% bonus for each level any cew member is
above the required minimum. So, if you have a lab that requires a level 1
scientist, having a level 5 scientist in it will give you a 40% bonus. The
effect accumulates: having two level 5 scientists will give you 80%.