Laboratory / Harvester efficiency increases with crew level (#204)
* Increase research speed in Laboratory if high level crew members are in it.

The laboratory will now give a 10% bonus for each level any cew member is
above the required minimum. So, if you have a lab that requires a level 1
scientist, having a level 5 scientist in it will give you a 40% bonus. The
effect accumulates: having two level 5 scientists will give you 80%.

* Same goes for engineers and harvesters.

Having an engineers on board will make a harvester 20% more effective.
This effect increaes with the level of the engineers.

* Having any engineer should provide a bonus.

This is an incentive to have your minig bases crewed.

* Don't overdo it: effect clamped to max 70%

* Put the crew level gain values into the config file

* Final tweaks to harvester speed-up.

- any engineer should increase the efficiency, not just high level ones.
- just having an engineer will provide a nice bump to the harvester speed,
  no matter how low the level.

* Microsofts wheels are just as round as all the others. But they tend to woblle a bit more.