Fix SSPX 2.5m Greenhouse
This 2.5 SSPX greenhouse was producing just over 3kg of food every 10 or so minutes.

Kerbals were having to eat faster and faster to keep up.  Linus wanted to know how it happened, Gus was concerned that Jeb was spending more time eating than piloting and Wernher was considering how this could be utilised as a propulsion source, turning this part into a Vomit Comet.  As much as Gene was taken with a re-purposing a malfunctioning part to complete a mission, Mortimer reminded him that the PAID for engines still worked fine.  Someone through a patch over the part and all was well.

Reduced the crop rate from the accidental high mentioned above to the expected 11 day turn around based 1/18 of the SSPX 3.75m time and crop rate.