Another alteration to the Partmodule Upgrade thing in Default Profile (#311)
* Update Default.cfg

Allows for other Profiles to not update slots or do other things.

* Update CHANGELOG.md

* Update Default.cfg again

Apparently my MM Fu is not great.  and MM was complaining bitterly about the :FOR[Kerbalism]
I quote the esteemed blowfish: 
"It's complaining about :FOR[Kerbalism] - the :NEEDS is fine.  Non-patch nodes can't have a pass - they're all inserted at the beginning."
So this removes the offensive term.

* Update CHANGELOG.md