Reverted the no-flow restrictions introduced by 025c1cb822bf8d58eb99776e123dbb9ec2ca0fee
Process simulation will now be done vessel-wide without a relation to a part. This is how
it was before that commit. I somewhat get why the per-part simulation was done, however I
don't think that it should ever have been introduced. From what I gathered it was to support
one mod, but turned the entire simulation upside down and caused a series of issues in
seemingly unrelated places. And why a pseudo-resource needed to be no-flow eludes me.

If that mod (near future electrical) needs special handling, it has to be done differently.