Revert part processing introduced with PR #280 (#371)
* Reverted the no-flow restrictions introduced by 025c1cb822bf8d58eb99776e123dbb9ec2ca0fee

Process simulation will now be done vessel-wide without a relation to a part. This is how
it was before that commit. I somewhat get why the per-part simulation was done, however I
don't think that it should ever have been introduced. From what I gathered it was to support
one mod, but turned the entire simulation upside down and caused a series of issues in
seemingly unrelated places. And why a pseudo-resource needed to be no-flow eludes me.

If that mod (near future electrical) needs special handling, it has to be done differently,
maybe by adding new part modules with some sort of special handling.

This fixes #333
This probably breaks parts of Near Future Electrics in some way
This might be related to #356, but not sure

* and a changelog to go with it.