I lied
I thought the ec rates and duration would be affected by the patch inside NewScience.cfg. which multiplies ec cost by 5 and halves the duration

Apparently, that's not the case, and the experiments in the group have their true value from the tweakable applied ingame. Why? beyond me. Will probably need a revisit, and some order tweaks so it doesn;t break in the future. Functional for now.