Combining multiple antennae no longer sums up the total data rate.
We use the geometric mean of the rates instead.

What this means in practice: if you combine 2 or more antennae of
the same type, you will increase the total communication range of
the vessel and at the same time decrease the loss of data rate with
distance. When you compare a vessel with one antenna with another
vessel that has two antennae, they will be able to transmit (almost)
equally fast over short distances. But at long range, the vessel
with 2 antennae will have an advantage.

This also means that you will want to combine antennae of same or
similar capabilities only. If you combine one very fast antenna with
one that is very weak, you will sacrifice almost all of the speed
benefits you get for the fast antenna for the added range benefit
of the weak antenna.