PLS DON'T GO BOOM.
Reordered everything.
This should make patching mods not require juggling PASSes or head scratches as to where i should patch them.

I did some quick tests, stuff works as intended, so there's no functional changes compared to how it was yesterday, but now i don't really care about when or where other mods patch their stuff.
Misc.cfg is there more as an explanation as to what the fuck i'm doing (you know, for bus insurance, in case someone wants to modify it in the future). It's still required, so don't delete it.

I'm sure i missed some shit, due to having to re-do the order of every single patch, but it's not throwing errors, so it may be slightly more tricky to spot. Mod Support  suffered the same treatment, but I did not test them. Too tired atm to do it. Will do tomorrow, and i'll get back to patching.