Update Resources.cfg
quote @sswelm

According to the resource file , MoistAtmosphere supposededly has a density of 0.000804. I notice that is very close to the density of Water which is 0.001. Saturated water vapor at STP has a true density of 4.85 g/m3 at 273 K , which translates into a density of 0.00000485 which is 165.77 times less than currently!!