Be graceful with multiple hard drives in one part, even if it is not supported
Having more than one hard drive part module in one part is not supported,
but still happens often enough with broken config setups that I think it
should be handled by Kerbalism in a somewhat cooperative way.

If a second hard drive part module is detected on any part, the capacity
of the second drive is added to the capacity of the first drive and the
second drive is then disabled (isEnabled = false).

Fixes #479