Virtual resources implementation.
This introduce a IResource interface to be used in place of the VesselResource objets for the most common usages. Can be casted to VesselResource  to get the extra information if needed.

VirtualResource also implement IResource and is accessed and used in the same way as a regular resource. It is accessed trough ResourceCache and can use Produce(), Consume() as well as being an input or output in a Recipe. To create a VirtualResource, just use one of the GetResource method with a name that doesn't exist in the KSP resources.

It is meant to be used in the following cases :
- replacement for processes pseudoresources (note : this will require migrating the planner resource sim before that can happen)
- checking the output of a resource consumer, by creating a Recipe with the consumed resource as input and the virtual resource as output
- eventually they could be used to replace the habitat pseudo resources but well... don't touch habitat.