When setting a max warp rate, set it according to desired warp speed, not step index.
There are mods (looking at you, RO) that change the warp speed steps, so when warp
step n.5 is 1000x in stock, it can be much higher with those mods. Enforcing that
wrong warp step in a loop makes it impossible for the player to get out of time
warp, because it's always reset.