Allow partial recording of stock science experiment data.
Partial recording will fill up all drives on a vessel with the experiment data and discard the rest.
This is better than recording nothing at all, this way unsupported experiments can be done in full
by repeating them in situ until all scientific value was transmitted.

Before that, the experiment could not be done at all.