New formula to calculate the shadow period for elliptical orbits.
The new metod returns consistent results for vessels on elliptical
orbits, it no longer depends on the vessel altitude at the time of
method call. But it still has limitations:

The method gives the longest amount of time spent in darkness, which for
some orbits (e.g.polar orbits), will only be experienced periodically.
It ignores inclination, the argument of periapsis and the beta angle
(the angle the sun is in relation to the orbit). Even polar orbits above the
day/night terminator line (which would be in the sun all the time) will
be treated as orbits passing the nadir at apoapsis.