err on the light side
if more than half the orbit duration is in shadow, invert
the result. this will be wrong when the apoapsis is near nadir.
this is just... WRONG, I know, but it is wrong in a good way.
f.i. a very eccentric orbit with the apoapsis above one of the
poles won't be assumed to spend most of the time in shadow.