ResourceSimulator: Use real calculation for simulated Process_harvester
Pulled out a public static method in Harvester so the simulator can
share some code with the real module, and changed dump to false to match
the real code.

Also needed to make the full crew list available as a member on
VesselAnalyzer in order to properly calculate the engineer-levels bonus.