Increased the starting altitude of the inner radiation belt.
Due to the offset of the geomagnetic center, the inner radiation belt reached into the upper atmosphere of kerbin. Increased the starting altitude of the inner belt a bit, also in the hope that it will now be possible to find a mun return trajectory that never enters the inner radiation belt at all.