Makes Kerbalism possibly compatible with future KSP 1.x versions without intervention.
With this, chances are that Kerbalism versions will continue to work with newer KSP 1.x
versions - until now, the shader folder had to be copied to a folder that mateched the
KSP version numbers. Since KSP 1.5 (and possibly before), it has not been necessary to
recompile the shaders. Kerbalism contained the same set of shader files for 1.5-1.9.
Chances are that future versions of KSP will still continue to work with these. To avoid
the need to keep multiple copies of the same files, or manually rename the shader folder
after a KSP update, use the default shader folder for all versions. If needed, this can be
changed for future versions if they ever should require a new set of shaders.