Prepare for major version change (3.x -> 4.0), drop support for KSP before 1.8
1. On the major version change

The current development will add a list of new features and at the same time
remove some of the major bugs in Kerbalism - tasks that entail changes in large
parts of the code base as well as game configuration. Some of those changes
will break compatibility with existing saves. We tried to avoid that, but as it
turned out, the amount of work needed to maintain backward compatibility more
than doubled the work, and would include having to carry a number of quirks and
workarounds for a long time, which in turn would also increase the workload in
the forseeable future. It would be a maintenance nightmare, especially for a
not-for-profit project.

2. Dropping support for KSP older than 1.8

With the update from KSP 1.7 to 1.8, Squad also changed the Unity version as
well as the supporting framework that the game runs on. The newer framework
comes with a lot of interesting features and possible performance improvements,
things that Kerbalism was not able to use to the full extent possible because
the code base also had to run on the older frameworks for the older game
versions. There were a couple of source code conditionals, meaning that some
routines had to be implemented twice (one for KSP < 1.8, one for KSP >= 1.8)
which means that, at least to a small degree, maintaining compatibility with
older game versions increased the workload.