Redefinition of science config.
The Kerbalism science configuration has been extended. It is now possible to
specifiy desired data size and experiment duration directly inside the
EXPERIMENT_DEFINITION node, with multiple experiment module configuration
variations when needed. This simplifies the science configuration process a lot.

Example:

EXPERIMENT_DEFINITION:NEEDS[KerbalismScience]
{
  id = kerbalism_BIRDIE
  title = #kerbalism-experiment-birdie-title

  KERBALISM_EXPERIMENT
  {
    Description = #kerbalism-experiment-birdie-desc
    DataSize = 5500
    SciencePoints = 50

    MODULE_DEFINITION
    {
      // name = kerbalism_BIRDIE // optional, if not specified will use experiment id
      Duration = 150d
      RequiredEC = 1.5
      Requirements = CrewMin:2
      CrewRequirement = Scientist:2
    }

    Situation = InSpaceLow
    Situation = InSpaceHigh
    BodyNotAllowed = HomeBody
  }
}

@PART[...]
{
  MODULE
  {
    name = ModuleKsmExperiment
    id = kerbalism_BIRDIE
  }
}

The existing science configuration was migrated to the new definition, and cleaned
up in the process. A couple of experiments were removed, the science groups were
replaced by dedicated science parts and new science configurations for the processing
lab and crewed science.

There are a couple of broken science configurations, but those will produce MM errors
and can be fixed later.

Simplified the ModuleKsmExperiment part module (removed setup and reset crew
requirements), renamed hide_when_unavailable to hide_when_invalid (will hide the UI
in the PAW while in invalid situations, this is used on the MPL).

STAKE now works differently: it is a 6 year experiment yielding a total of 1500 science
points, but you have to "buy" that by providing the extremely expensive "STAKE granulate"
resource to the experiment. It only runs on the MPL in Kerbin orbit, and the amount that
can be stored in the science box doesn't last for more than ~200 days, so you will have to
run resupply missions. A science box of resupply costs about 180.000 in funds.