More substep sim fixes and tweaks : - Full substeps recompute now only required on SOI change if on rails - Continously recompute all substeps on loaded vessels instead of checking for orbit parameters changes (too unreliable). - Cleaned a few things - The mini profiler now show sim time consumed / missed. The time missed is the critical metric to determine if the vessel count and/or max timewarp rate are too high for the sim thread processing capacity.
After some tests, I think the best way to handle higher than stock timewarp rates is to have an adaptative (or at least configurable) substep interval. In the case of bigger scale systems (RSS...) it make sense to increase the interval.

In stock, a LKO orbit is ~30 min, meaning 30 sampling points per orbit at the current interval of 1 minute.
In RSS, a LEO orbit is ~90 minutes, so it wouldn't feel out of place to put the interval at 3 minutes, resulting in the same amount of sampling points per orbit.

It seems that in general, it's hard to get a an orbital period that is less than 1 hour, no matter the body size. KSP orbital periods are very low because of the completely unrealistic body densities used to get real range gravity values while using 1/10 the real body sizes.