Self consuming processes + a few fixes
In the modifier formula, you can use a virtual resource called by default "process name" + "Process"
You can change that name with "pseudoResourceName = X".
Note that this name must be unique and different from any regular KSP resource name.
The default modifier is : ResHandler.GetResource(pseudoResourceName).Amount
The pseudo resource Capacity is set to the sum of active process controllers capacity
The pseudo resource Amount is the Capacity scaled by the user utilization setting

If you set UseCapacityResource = true, a local virtual resource named by default
"process name" + "Capacity" will added to the part by each process controller.
You can change that name with "capacityResourceName = X" (must also be unique).
The inital amount and capacity is set to the process controller capacity.
Contrary to the process pseudo-resource, that capacity resource can be used
in the inputs / outputs of a process, as well as in a process modifier.
This allow to create self consuming processes, see the RTG and nonRegenerativeScrubber
for examples.

By default, the capacity resource won't be visible. This can be changed by providing
a VIRTUAL_RESOURCE definition of type = PartResource for it in the profile.