Update SSPX.cfg
Hi! I added support for the new greenhouses, as well as the different bonuses (panorama, exercise, plants) for the new parts of SSPX. All the areas and volumes have been calculated importing the models in blender, and the grow rates are based on the the original Kerbalism Greenhouse. Todo: the science stuff.